+++ src/06-entities/User/index.ts
@@ -1,2 +1,2 @@
-export type { User, UserInitialState } from './types/user.types';
 export { userActions, userReducer } from './model/slice/user.slice';
+export type { User, UserInitialState } from './types/user.types';
