+++ src/07-shared/lib/components/index.ts
@@ -1,0 +1,14 @@
+import { AsyncReducerLoader, ReducerList } from './AsynReducerLoader/AsyncReducerLoader';
+import { ErrorBoundary } from './ErrorBoundary';
+import { Portal } from './Portal/Portal';
+import { Theme, ThemeProvider, useTheme } from './ThemeProvider';
+
+export {
+    AsyncReducerLoader,
+    ReducerList,
+    ErrorBoundary,
+    Portal,
+    Theme,
+    ThemeProvider,
+    useTheme,
+};
