+++ src/07-shared/ui/AppNavLink/AppNavLink.test.tsx
@@ -1,6 +1,6 @@
-import { withMemoryRouter } from '07-shared/lib/testsHelpers/withMemoryRouter';
 import { render, screen } from '@testing-library/react';
-import { AppNavLink } from '07-shared/ui/AppNavLink/AppNavLink';
+import { withMemoryRouter } from '07-shared/lib/testsHelpers/withMemoryRouter';
+import { AppNavLink } from './AppNavLink';
 
 describe('AppNavLink', () => {
     test('AppNavLink render with active path', () => {
