+++ src/07-shared/ui/Button/Button.tsx
@@ -1,5 +1,5 @@
 import { ButtonHTMLAttributes, memo, ReactNode } from 'react';
-import { classNames } from '../../lib/classNames/classNames';
+import { classNames } from '07-shared/lib/classNames/classNames';
 import classes from './Button.module.scss';
 
 export type ButtonType = 'clear' | 'outline' | 'primary' | 'secondary';
