+++ src/07-shared/ui/Input/Input.tsx
@@ -1,11 +1,11 @@
 import {
     ChangeEvent, InputHTMLAttributes, memo, ReactNode, useState,
 } from 'react';
-import classes from './Input.module.scss';
-import Cross from '../../assets/icons/cross.svg';
-import EyeOpen from '../../assets/icons/eye-open.svg';
-import EyeClose from '../../assets/icons/eye-closed.svg';
+import Cross from '07-shared/assets/icons/cross.svg';
+import EyeOpen from '07-shared/assets/icons/eye-open.svg';
+import EyeClose from '07-shared/assets/icons/eye-closed.svg';
 import { classNames } from '../../lib/classNames/classNames';
+import classes from './Input.module.scss';
 
 type InputTypes = 'primary' | 'clear';
 type LabelPosition = 'top' | 'left';
